an advance in py3

+++ dataanalysis/hashtools.py
@@ -74,7 +74,7 @@ def hashe_list_objects(hashe):
 
 def remove_repeating_stacks(input_stack):
     exclude_mask=[False]*len(input_stack)
-    for stack_length in range(1,len(input_stack)/2):
+    for stack_length in range(1,int(len(input_stack)/2)):
         for stack_start in range(0,len(input_stack)-stack_length):
             if input_stack[stack_start:stack_start+stack_length] == input_stack[stack_start+stack_length:stack_start+stack_length+stack_length]:
                 log("found repetition of ",stack_start,stack_length,":",input_stack[stack_start:stack_start+stack_length*2],level="top")
